Move `cpt` to `d3compat` module
