+++ src/charts/html-legend.js
@@ -1,6 +1,7 @@
 import {select} from 'd3-selection';
 
-import {cpt, pluck, utils} from '../core/utils';
+import {pluck, utils} from '../core/utils';
+import {cpt} from '../core/d3comapt';
 import {constants} from '../core/constants';
 
 /**
