+++ src/charts/select-menu.js
@@ -1,7 +1,7 @@
 import {events} from '../core/events';
 import {BaseMixin} from '../base/base-mixin';
 import {logger} from '../core/logger';
-import {cpt} from '../core/utils';
+import {cpt} from '../core/d3comapt';
 
 const SELECT_CSS_CLASS = 'dc-select-menu';
 const OPTION_CSS_CLASS = 'dc-select-option';
